feat: refactor auth to modules structure
