+++ src/app/(auth)/signup/page.tsx
@@ -1,22 +1,5 @@
-import { GalleryVerticalEnd } from "lucide-react";
+import SignUpPage from "@/modules/auth/signup.page";
 
-import { SignupForm } from "../_components/signup-form";
-
-export default function SignUpPage() {
-    return (
-        <div className="bg-muted flex min-h-svh flex-col items-center justify-center gap-6 p-6 md:p-10">
-            <div className="flex w-full max-w-sm flex-col gap-6">
-                <a
-                    href="#"
-                    className="flex items-center gap-2 self-center font-medium"
-                >
-                    <div className="bg-primary text-primary-foreground flex size-6 items-center justify-center rounded-md">
-                        <GalleryVerticalEnd className="size-4" />
-                    </div>
-                    Acme Inc.
-                </a>
-                <SignupForm />
-            </div>
-        </div>
-    );
+export default async function Page() {
+    return <SignUpPage />;
 }
