+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation";
-import { getSession } from "@/lib/auth-utils";
+import { getSession } from "@/modules/auth/utils/auth-utils";
 
 export default async function HomePage() {
     const session = await getSession();
