+++ src/modules/auth/auth.route.ts
@@ -1,0 +1,6 @@
+const authRoutes = {
+    Login: "/login",
+    Signup: "/signup",
+} as const;
+
+export default authRoutes;
