+++ src/modules/auth/components/signup-form.tsx
@@ -7,7 +7,6 @@ import { useRouter } from "next/navigation";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
 import toast from "react-hot-toast";
-import { signIn } from "@/actions/auth.action";
 import { Button } from "@/components/ui/button";
 import {
     Card,
@@ -25,23 +24,25 @@ import {
     FormMessage,
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
-import { authClient } from "@/lib/auth-client";
 import { cn } from "@/lib/utils";
 import {
-    type SignInSchema,
-    signInSchema,
-} from "@/lib/validations/auth.validation";
+    type SignUpSchema,
+    signUpSchema,
+} from "@/modules/auth/models/auth.model";
+import { signUp } from "../actions/auth.action";
+import { authClient } from "../utils/auth-client";
 
-export function LoginForm({
+export function SignupForm({
     className,
     ...props
 }: React.ComponentProps<"div">) {
     const router = useRouter();
     const [isLoading, setIsLoading] = useState(false);
 
-    const form = useForm<SignInSchema>({
-        resolver: zodResolver(signInSchema),
+    const form = useForm<SignUpSchema>({
+        resolver: zodResolver(signUpSchema),
         defaultValues: {
+            username: "",
             email: "",
             password: "",
         },
@@ -54,9 +55,9 @@ export function LoginForm({
         });
     };
 
-    async function onSubmit(values: SignInSchema) {
+    async function onSubmit(values: SignUpSchema) {
         setIsLoading(true);
-        const { success, message } = await signIn(values);
+        const { success, message } = await signUp(values);
 
         if (success) {
             toast.success(message.toString());
@@ -71,9 +72,9 @@ export function LoginForm({
         <div className={cn("flex flex-col gap-6", className)} {...props}>
             <Card>
                 <CardHeader className="text-center">
-                    <CardTitle className="text-xl">Welcome back</CardTitle>
+                    <CardTitle className="text-xl">Welcome</CardTitle>
                     <CardDescription>
-                        Login with your Google account
+                        Sign up with your Google account
                     </CardDescription>
                 </CardHeader>
                 <CardContent>
@@ -98,7 +99,7 @@ export function LoginForm({
                                             fill="currentColor"
                                         />
                                     </svg>
-                                    Login with Google
+                                    Sign up with Google
                                 </Button>
                                 <div className="after:border-border relative text-center text-sm after:absolute after:inset-0 after:top-1/2 after:z-0 after:flex after:items-center after:border-t">
                                     <span className="bg-card text-muted-foreground relative z-10 px-2">
@@ -106,6 +107,22 @@ export function LoginForm({
                                     </span>
                                 </div>
                                 <div className="grid gap-6">
+                                    <FormField
+                                        control={form.control}
+                                        name="username"
+                                        render={({ field }) => (
+                                            <FormItem>
+                                                <FormLabel>Username</FormLabel>
+                                                <FormControl>
+                                                    <Input
+                                                        placeholder="johndoe"
+                                                        {...field}
+                                                    />
+                                                </FormControl>
+                                                <FormMessage />
+                                            </FormItem>
+                                        )}
+                                    />
                                     <FormField
                                         control={form.control}
                                         name="email"
@@ -122,33 +139,23 @@ export function LoginForm({
                                             </FormItem>
                                         )}
                                     />
-                                    <div className="flex flex-col gap-2">
-                                        <FormField
-                                            control={form.control}
-                                            name="password"
-                                            render={({ field }) => (
-                                                <FormItem>
-                                                    <FormLabel>
-                                                        Password
-                                                    </FormLabel>
-                                                    <FormControl>
-                                                        <Input
-                                                            placeholder="*********"
-                                                            {...field}
-                                                            type="password"
-                                                        />
-                                                    </FormControl>
-                                                    <FormMessage />
-                                                </FormItem>
-                                            )}
-                                        />
-                                        <a
-                                            href="#"
-                                            className="ml-auto text-sm underline-offset-4 hover:underline"
-                                        >
-                                            Forgot your password?
-                                        </a>
-                                    </div>
+                                    <FormField
+                                        control={form.control}
+                                        name="password"
+                                        render={({ field }) => (
+                                            <FormItem>
+                                                <FormLabel>Password</FormLabel>
+                                                <FormControl>
+                                                    <Input
+                                                        placeholder="*********"
+                                                        {...field}
+                                                        type="password"
+                                                    />
+                                                </FormControl>
+                                                <FormMessage />
+                                            </FormItem>
+                                        )}
+                                    />
                                     <Button
                                         type="submit"
                                         className="w-full"
@@ -160,17 +167,17 @@ export function LoginForm({
                                                 Loading...
                                             </>
                                         ) : (
-                                            "Login"
+                                            "Sign Up"
                                         )}
                                     </Button>
                                 </div>
                                 <div className="text-center text-sm">
-                                    Don&apos;t have an account?{" "}
+                                    Already have an account?{" "}
                                     <Link
-                                        href="/signup"
+                                        href="/login"
                                         className="underline underline-offset-4"
                                     >
-                                        Sign up
+                                        Sign in
                                     </Link>
                                 </div>
                             </div>
