+++ src/modules/auth/models/auth.model.ts
