+++ src/modules/auth/models/user.model.ts
@@ -1,0 +1,5 @@
+export interface AuthUser {
+    id: string;
+    name: string;
+    email: string;
+}
